+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added GenCast 1 degree Mini model
+- Added SamudrACE Coupled Climate Model
 
 ### Changed